chore(bot): move webhook processors into common package
