+++ processors/processors_test.go
@@ -1,0 +1,13 @@
+package processors
+
+func intVal(i int) *int {
+	return &i
+}
+
+func stringVal(s string) *string {
+	return &s
+}
+
+func boolVal(b bool) *bool {
+	return &b
+}
