+++ processors/pull_request_review.go
@@ -1,0 +1,14 @@
+package processors
+
+import "github.com/google/go-github/github"
+
+func PullRequestReviewEvent(client *github.Client, input <-chan *github.PullRequestReviewEvent) <-chan *github.PullRequest {
+	ret := make(chan *github.PullRequest)
+	go func() {
+		for evt := range input {
+			ret <- evt.PullRequest
+		}
+		close(ret)
+	}()
+	return ret
+}
