+++ processors/types.go
@@ -1,0 +1,8 @@
+package processors
+
+type Repository struct {
+	Owner    string
+	Name     string
+	Mainline string
+	Cache    WorkerCache
+}
